inopsy-agent/perf/requirements.txt
catboost>=1.0.6
lightgbm>=3.3.2
matplotlib>=3.4.3
numpy>=1.19.5
openpyxl>=3.0.9
pandas>=1.3.4
PyYAML>=6.0
scikit_learn>=1.1.2
seaborn>=0.12.0
xgboost>=1.6.2